+++ plugins/org.komodo.relational/src/org/komodo/relational/model/internal/ParameterImpl.java
@@ -228,8 +228,8 @@ public final class ParameterImpl extends RelationalChildRestrictedObject impleme
      * @see org.komodo.relational.model.Parameter#getPrecision(org.komodo.spi.repository.Repository.UnitOfWork)
      */
     @Override
-    public int getPrecision( final UnitOfWork uow ) throws KException {
-        final Integer value = getObjectProperty( uow, PropertyValueType.INTEGER, "getPrecision", //$NON-NLS-1$
+    public long getPrecision( final UnitOfWork uow ) throws KException {
+        final Long value = getObjectProperty( uow, PropertyValueType.LONG, "getPrecision", //$NON-NLS-1$
                                                  StandardDdlLexicon.DATATYPE_PRECISION );
 
         if (value == null) {
@@ -292,8 +292,8 @@ public final class ParameterImpl extends RelationalChildRestrictedObject impleme
      * @see org.komodo.relational.model.Parameter#getScale(org.komodo.spi.repository.Repository.UnitOfWork)
      */
     @Override
-    public int getScale( final UnitOfWork uow ) throws KException {
-        final Integer value = getObjectProperty( uow, PropertyValueType.INTEGER, "getScale", //$NON-NLS-1$
+    public long getScale( final UnitOfWork uow ) throws KException {
+        final Long value = getObjectProperty( uow, PropertyValueType.LONG, "getScale", //$NON-NLS-1$
                                                  StandardDdlLexicon.DATATYPE_SCALE );
 
         if (value == null) {
@@ -493,11 +493,11 @@ public final class ParameterImpl extends RelationalChildRestrictedObject impleme
     /**
      * {@inheritDoc}
      *
-     * @see org.komodo.relational.model.Parameter#setPrecision(org.komodo.spi.repository.Repository.UnitOfWork, int)
+     * @see org.komodo.relational.model.Parameter#setPrecision(org.komodo.spi.repository.Repository.UnitOfWork, long)
      */
     @Override
     public void setPrecision( final UnitOfWork uow,
-                              final int newPrecision ) throws KException {
+                              final long newPrecision ) throws KException {
         setObjectProperty( uow, "setPrecision", StandardDdlLexicon.DATATYPE_PRECISION, newPrecision ); //$NON-NLS-1$
     }
 
@@ -531,11 +531,11 @@ public final class ParameterImpl extends RelationalChildRestrictedObject impleme
     /**
      * {@inheritDoc}
      *
-     * @see org.komodo.relational.model.Parameter#setScale(org.komodo.spi.repository.Repository.UnitOfWork, int)
+     * @see org.komodo.relational.model.Parameter#setScale(org.komodo.spi.repository.Repository.UnitOfWork, long)
      */
     @Override
     public void setScale( final UnitOfWork uow,
-                          final int newScale ) throws KException {
+                          final long newScale ) throws KException {
         setObjectProperty( uow, "setScale", StandardDdlLexicon.DATATYPE_SCALE, newScale ); //$NON-NLS-1$
     }
